Semantic Analyzer: Added overloading based on return type.
If two functions are both equal contenders for being resolved to, only then look at the return type. The return type of the function has to match *exactly* with the type it is expected to return, so no implicitly compatible stuff. This is so the intended overload is always used.